requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
qrcode==7.4.2
Pillow==10.1.0
python-socketio==5.10.0
python-engineio>=4.8.0